requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.1.0
autopep8==1.5.6
bcrypt==3.2.0
bigquery==0.0.12
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dacktool==0.0.7
dbstream==0.1.0
dotmap==1.3.23
google-api-core==1.26.3
google-api-python-client==1.7.11
google-auth==1.30.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.14.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googleauthentication==0.0.15
grpcio==1.37.0
httplib2==0.19.1
idna==3.1
libcst==0.3.18
libxml2-python==2.9.10
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.15.8
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
PyYAML==5.4.1
requests-oauthlib==1.3.0
rsa==4.7.2
sshtunnel==0.1.5
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
yarl==1.6.3
